- added postman test: Endpoint test.postman...

+++ day3BE/rest/src/main/java/com/example/rest/service/UserService.java
@@ -1,5 +1,6 @@
 package com.example.rest.service;
 
+import com.example.rest.exception.UserNotFoundException;
 import com.example.rest.model.User;
 import com.example.rest.repository.IUserRepository;
 import org.springframework.data.domain.Page;
@@ -30,23 +31,25 @@ public class UserService {
     }
 
     public void updateUser(Integer id, String username, String email) {
-        User user = new User();
-        user.setId(id);
+        User user = userRepository.findById(id)
+                .orElseThrow(() -> new UserNotFoundException("User with ID " + id + " not found"));
+
         user.setUsername(username);
         user.setEmail(email);
-        User existingUser = userRepository.findById(id).orElseThrow(() -> new RuntimeException("User not found"));
-        existingUser.setUsername(user.getUsername());
-        existingUser.setEmail(user.getEmail());
-        userRepository.save(existingUser);
+        userRepository.save(user);
     }
 
     public void updateEmailForUser(Integer id, String email) {
-        User user = userRepository.findById(id).orElseThrow(() -> new RuntimeException("User not found"));
+        User user = userRepository.findById(id)
+                .orElseThrow(() -> new UserNotFoundException("User with ID " + id + " not found"));
+
         user.setEmail(email);
         userRepository.save(user);
     }
 
     public void deleteUser(Integer id) {
+        if (!userRepository.existsById(id)) {
+            throw new UserNotFoundException("User with ID " + id + " not found");
+        }
         userRepository.deleteById(id);
-    }
-}
+    }}
